BetterTSE-main/requirements.txt
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
openai>=1.0.0
scipy>=1.11.0